Move more View related files into the View subfolder
